pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vsanimesr"
version = "1.0.0"
description = "AnimeSR function for VapourSynth"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name = "HolyWu", email = "holywu@gmail.com"}]
keywords = ["AnimeSR", "VapourSynth"]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Topic :: Multimedia :: Video"
]
dependencies = [
  "numpy",
  "torch>=1.13.0",
  "torch-tensorrt-fx-only>=1.3.0",
  "VapourSynth>=55"
]

[project.urls]
"Homepage" = "https://github.com/HolyWu/vs-animesr"
"Bug Tracker" = "https://github.com/HolyWu/vs-animesr/issues"
